requirements.txt
flask==2.2.5
python-dotenv==1.0.0
openai==1.70.0
pytest==7.4.0
werkzeug==2.2.3
pymongo==4.5.0